space-app-backend/weather-backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.2
pandas==2.2.3
xarray==2024.9.0
netCDF4==1.7.2
